requirements.txt
django
djangorestframework
pygments
httpie